Added clearer structure to KillerTask
Now basically done except for UI-side implementation and retrieval of exclusion list

+++ app/src/main/java/sg/edu/nushigh/arp/backgroundprioritizer/KillerTask.java
@@ -3,18 +3,33 @@ package sg.edu.nushigh.arp.backgroundprioritizer;
 import android.content.Context;
 import android.os.AsyncTask;
 
-public class KillerTask extends AsyncTask<Context, Void, Void> {
-    // TODO we can have better type arguments, will use as we develop
+import java.util.ArrayList;
+import java.util.Set;
+import java.util.TreeSet;
+
+public class KillerTask extends AsyncTask<Void, Void, ArrayList<Utilities.ProcessUsageData>> {
+
+    private final Context c;
+    private final Set<String> exclusionList = new TreeSet<>();
+
+    KillerTask(Context c){
+        super();
+        this.c = c;
+        // TODO add relevant UI-related fields here
+        // TODO get exclusion list from some persistent location
+    }
 
     @Override
     protected void onPreExecute(){
         // what do we need, conditions checking?
+        // TODO this runs on UI thread, so we can set some loading messages
     }
 
     @Override
-    protected Void doInBackground(Context... params) {
+    protected ArrayList<Utilities.ProcessUsageData> doInBackground(Void... params) {
+        ArrayList<Utilities.ProcessUsageData> killed = new ArrayList<>();
         Utilities.ProcessUsageData[] data = Utilities.taskList();
-        String foregroundActivityPackageName = Utilities.getForegroundActivity(params[0]);
+        String foregroundActivityPackageName = Utilities.getForegroundActivity(c);
         for(Utilities.ProcessUsageData d: data){
             // skip system processes for stability
             if(d.getUid().equals("system"))
@@ -23,17 +38,23 @@ public class KillerTask extends AsyncTask<Context, Void, Void> {
                 continue;
             if(d.getUid().equals("shell"))
                 continue;
+            // skip processes in exclusion list
+            if(exclusionList.contains(d.getName()))
+                continue;
             // skip foreground process (always backgroundprioritizer for now in this implementation)
             if(d.getName().equals(foregroundActivityPackageName))
                 continue;
             // if none of the above, we kill the process
             Utilities.executeSuperCommand("kill " + d.getPid());
+            killed.add(d);
         }
-        return null;
+        return killed;
     }
 
     @Override
-    protected void onPostExecute(Void result){
-        // a list of killed tasks?
+    protected void onPostExecute(ArrayList<Utilities.ProcessUsageData> result){
+        // result is list of killed tasks
+        // TODO this runs on UI thread , display whatever UI-side results we want
+
     }
 }
